consolidate import_installed_app_flow()
